Refactoring Fixing group management when api key was reset

+++ includes/class-wp-compare.php
@@ -239,7 +239,6 @@ class Wp_Compare {
 			'group_id'		=> $group_id,
 			'api_key'		=> $api_key
 		);
-
 		return mm_api( $args );
 	}
 
@@ -253,13 +252,14 @@ class Wp_Compare {
 		);
 
 		$api_key = mm_api( $args );
+		if( isset( $api_key['status'] ) && $api_key['status'] == 'success' ) {
 
-		update_option( 'wpcompare_api_key', $api_key );
-		delete_option( 'wpcompare_group_id' );
-		delete_option( 'wpcompare_monitoring_group_id' );
+            update_option('wpcompare_api_key', $api_key['api_key']);
+            delete_option('wpcompare_group_id');
+            delete_option('wpcompare_monitoring_group_id');
 
-		$this->create_group( $api_key );
-
+            $this->create_group($api_key['api_key']);
+        }
 		return $api_key;
 	}
 
@@ -291,16 +291,30 @@ class Wp_Compare {
 			return $api_key;
 		} else
 			return false;
-
 	}
 
 	function get_api_key_form( $api_key = false ) {
 
-		return '<form action="/wp-admin/admin.php?page=wp-compare&tab=settings" method="post">
-				API Key
-				<input type="text" name="api-key" value="' . $api_key .  '">
-				<input type="submit" value="Save" class="button">
-			</form>';
+
+	    if( $api_key ) {
+            $output = '<form action="/wp-admin/admin.php?page=wp-compare&tab=settings" method="post" 
+                        onsubmit="return confirm(\'Do you really want to reset the API key?\nYour settings will get lost.\');">
+                        <input type="hidden" name="action" value="reset_api_key">
+                        <h2>API Key</h2>
+                        <p>Your API key: <strong>' . $api_key . '</strong></p>
+                        <input type="submit" value="Reset API key" class="button"><br>
+                        <p><strong>ATTENTION: With resetting the API key, all settings get lost and 
+                        the monitoring won\'t be continued!</strong></p>';
+        } else {
+	        $output = '<form action="/wp-admin/admin.php?page=wp-compare&tab=settings" method="post">
+                        <input type="hidden" name="action" value="save_api_key">
+                        <h2>You already have an API key?</h2>
+                        <p>Enter your API key here and start comparing.</p>
+                        <input type="text" name="api-key" value="' . $api_key .  '">
+                        <input type="submit" value="Save" class="button">';
+        }
+	    $output .= '</form>';
+		return  $output;
 	}
 
 	function verify_api_key( $api_key ) {
@@ -324,9 +338,10 @@ class Wp_Compare {
 	function create_group( $api_key ) {
 		// Create group if it doesn't exist yet
 		$args = array(
-			'action'	=> 'add_group',
-			'domain'	=> $_SERVER['SERVER_NAME'],
-			'api_key'	=> $api_key
+			'action'	    => 'add_group',
+			'domain'	    => $_SERVER['SERVER_NAME'],
+			'website_group' => 1,
+			'api_key'	    => $api_key
 		);
 
 		$group = mm_api( $args );
@@ -337,6 +352,145 @@ class Wp_Compare {
 		update_option( 'wpcompare_group_id', $manual_group_id );
 		update_option( 'wpcompare_monitoring_group_id', $monitoring_group_id );
 	}
+
+	function delete_group( $group_id, $api_key ) {
+	    $args = array(
+	        'action'    => 'delete_group',
+            'group_id'  => $group_id,
+            'api_key'   => $api_key
+        );
+	    mm_api( $args );
+    }
+
+    function get_urls_of_group( $group_id ) {
+        $args = array(
+            'action'		=> 'get_group_urls',
+            'group_id'		=> $group_id
+        );
+        $group_urls = mm_api( $args );
+
+        $check_posts = array();
+        $amount_sc = 0;
+
+        foreach( $group_urls as $group_url ) {
+            // Create array with all active urls of group
+            $check_posts[] = (int)$group_url['wp_post_id'];
+            $check_desktop[$group_url['wp_post_id']] = $group_url['desktop'];
+            $check_mobile[$group_url['wp_post_id']] = $group_url['mobile'];
+
+            // Count amount of sc
+            if( $group_url['desktop'] )
+                $amount_sc++;
+            if( $group_url['mobile'] )
+                $amount_sc++;
+        }
+
+        return array(
+            'check_posts'   => $check_posts,
+            'check_desktop' => $check_desktop,
+            'check_mobile'  => $check_mobile,
+            'amount_sc'     => $amount_sc
+        );
+    }
+
+    function mm_get_url_settings( $group_id, $monitoring_group = false ) {
+
+        global $api_key;
+
+        $checks = $this->get_urls_of_group( $group_id );
+        // Get manual urls of group
+        /*$args = array(
+            'action'		=> 'get_group_urls',
+            'group_id'		=> $group_id,
+            'api_key'       => $api_key
+        );
+        $group_urls = mm_api( $args );
+
+        $check_posts = array();
+        foreach( $group_urls as $group_url ) {
+            $check_posts[] = (int)$group_url['wp_post_id'];
+            $check_desktop[$group_url['wp_post_id']] = $group_url['desktop'];
+            $check_mobile[$group_url['wp_post_id']] = $group_url['mobile'];
+        }*/
+
+        // Select URLS
+        if( $monitoring_group )
+            $tab = "monitoring-screenshots";
+        else
+            $tab = "take-screenshots";
+
+        echo '<form action="/wp-admin/admin.php?page=wp-compare&tab=' . $tab . '" method="post">';
+        echo '<input type="hidden" value="wp-compare" name="page">';
+        echo '<input type="hidden" value="post_urls" name="action">';
+        echo '<input type="hidden" value="' . $group_id . '" name="group_id">';
+
+        echo '<script>
+		function mmMarkRows( postId ) {
+			var checkBox = document.getElementById("active-" + postId);
+			var row = document.getElementById("post_id_" + postId );
+
+			if (checkBox.checked == true){
+				row.style.background = "#17b33147";
+			} else {
+				row.style.background = "#dc323247";
+			}
+		}
+	</script>';
+
+        $post_types = get_post_types();
+
+        foreach( $post_types as $post_type ) {
+
+            $posts = get_posts([
+                'post_type' => $post_type,
+                'post_status' => 'publish',
+                'numberposts' => -1,
+                'order'    => 'ASC'
+            ]);
+
+            if( $posts ) {
+                echo '<h2>' . ucfirst( $post_type ) . '</h2>';
+                echo '<table><tr><th>Active</th><th>Desktop</th><th>Mobile</th><th>Post Name</th><th>URL</th></tr>';
+                foreach( $posts as $post ) {
+                    $url = get_permalink( $post );
+
+                    // Check posts
+                    if( in_array( $post->ID, $checks['check_posts'] ) )
+                        $checked = 'checked';
+                    else
+                        $checked = '';
+
+                    // Check Desktop
+                    if( isset( $checks['check_desktop'][$post->ID] ) && $checks['check_desktop'][$post->ID] == 1 )
+                        $checked_desktop = 'checked';
+                    else
+                        $checked_desktop = '';
+
+                    // Check Mobile
+                    if( isset( $checks['check_mobile'][$post->ID] ) && $checks['check_mobile'][$post->ID] == 1 )
+                        $checked_mobile = 'checked';
+                    else
+                        $checked_mobile = '';
+
+                    echo '<tr id="post_id_' . $group_id . '-' . $post->ID . '">';
+                    echo '<td><input id="active-' . $group_id . '-' . $post->ID . '" onclick="mmMarkRows(\'' . $group_id . '-' . $post->ID . '\')" type="checkbox" name="pid-' . $post->ID . '" value="' . $post->ID . '" ' . $checked . '></td>';
+                    echo '<td><input type="hidden" value="0" name="desktop-' . $post->ID . '"><input type="checkbox" name="desktop-' . $post->ID . '" value="1" ' . $checked_desktop . '></td>';
+                    echo '<td><input type="hidden" value="0" name="mobile-' . $post->ID . '"><input type="checkbox" name="mobile-' . $post->ID . '" value="1" ' . $checked_mobile . '></td>';
+                    echo '<td>' . $post->post_title . '</td>';
+                    echo '<td><a href="' . $url . '" target="_blank">' . $url . '</a></td>';
+                    echo '</tr>';
+
+                    echo '<script>mmMarkRows(\'' . $group_id . '-' . $post->ID . '\'); </script>';
+
+
+                }
+                echo '</table>';
+
+            }
+        }
+        echo '<input class="button" type="submit" value="Save" style="margin-top: 30px">';
+        echo '</form>';
+    }
 
 	function get_no_account_page() {
 
@@ -391,6 +545,7 @@ class Wp_Compare {
 
 		$output =  '<div class="mm_wp_compare">
 		<h1>WP Compare</h1>
+		<h2>Create Free Account</h2>
 		<p>Create now your free account and <strong>100 compares</strong> for one month for free!</p>
 		<form id="frm_new_account" name="new_account" action="/wp-admin/admin.php?page=wp-compare&tab=settings" onsubmit="return mmValidateForm()" method="post">
 			<input type="hidden" name="action" value="create_free_account"><br>
@@ -399,7 +554,7 @@ class Wp_Compare {
 			<p><label>Email</label><input type="text" id="form_email" name="email"></p>
 			<input type="submit" value="Create free account" class="button">
 		</form>
-		<h2>You already have an API key?</h2>
+		<hr>
 		' . $this->get_api_key_form() . '
 		</div>';
 		return $output;
@@ -408,15 +563,10 @@ class Wp_Compare {
 
 function mm_api( $args ) {
 
-	//$url = 'https://comp.wpmike.com/v1/api.php';
 	$url = 'https://app.wpmike.com/v1/api.php';
-// 34.243.134.105
-	$api_key = get_option( 'wpcompare_api_key' );
-	//if( !$api_key )
-	//	return 'No API Key';
 
-	if( $api_key )
-		$args['api_key'] = $api_key;
+    if( !isset( $args['api_key'] ) )
+        $args['api_key'] = get_option( 'wpcompare_api_key' );
 
 	$ch = curl_init( $url );
 	//curl_setopt($ch, CURLOPT_USERPWD, "wpmike:letmein");
